feat: 评论组件增加点赞功能
- 新增点赞相关数据结构和接口
- 实现评论点赞/取消点赞功能
- 添加点赞数量和状态的缓存处理
- 更新数据库结构，增加点赞记录和汇总表
- 修改前端接口和组件以支持点赞功能

+++ vertx-app-service/vertx-framework/vertx-framework-component/src/main/java/com/onezol/vertx/framework/component/commet/model/dto/Comment.java
@@ -32,7 +32,7 @@ public class Comment extends BaseDTO {
     private User author;
 
     @Schema(description = "点赞数")
-    private Integer upvotes;
+    private Long upvotes;
 
     @Schema(description = "是否点赞")
     private Boolean upvoted;
